replace inih with c++ version, move influx and improve, seperate settings struct
too large commit, but things worked out that way.
Changed the origonal inih program to be a c++ version made by Github user jtilly.
Moved the large single struct for settings to a struct for each type of setting. Each struct has a constructor that uses inih.
Moved the influx function to a new file and changed to utilize the vector of sensors. Changed strings to c++ too.